move BufferedData out of BufferedQueryResult
