+++ src/include/duckdb/main/buffered_query_result.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/main/stream_query_result.hpp
+// duckdb/main/buffered_query_result.hpp
 //
 //
 //===----------------------------------------------------------------------===//
@@ -12,6 +12,7 @@
 #include "duckdb/main/query_result.hpp"
 #include "duckdb/parallel/interrupt.hpp"
 #include "duckdb/common/queue.hpp"
+#include "duckdb/main/buffered_data.hpp"
 
 namespace duckdb {
 
@@ -20,34 +21,6 @@ class ClientContextLock;
 class Executor;
 class MaterializedQueryResult;
 class PreparedStatementData;
-class BufferedQueryResult;
-
-class BufferedData {
-private:
-	static constexpr idx_t BUFFER_SIZE = 100000 / STANDARD_VECTOR_SIZE;
-
-public:
-	BufferedData(shared_ptr<ClientContext> context) : context(context) {
-	}
-
-public:
-	void Populate(unique_ptr<DataChunk> chunk);
-	unique_ptr<DataChunk> Fetch(BufferedQueryResult &result);
-	void AddToBacklog(InterruptState state);
-	void ReplenishBuffer(BufferedQueryResult &result);
-	bool BufferIsFull() const;
-
-private:
-	shared_ptr<ClientContext> context;
-	// Our handles to reschedule the blocked sink tasks
-	// TODO: min heap? (priority queue)
-	queue<InterruptState> blocked_sinks;
-	// Protect against populate/fetch race condition
-	mutex glock;
-	// Keep track of the size of the buffer to gauge when it should be repopulated
-	queue<unique_ptr<DataChunk>> buffered_chunks;
-	atomic<idx_t> buffered_chunks_count;
-};
 
 class BufferedQueryResult : public QueryResult {
 	friend class ClientContext;
